Move a few unsigned ints to NSUInteger, fix memory leak in NSIndexPath category.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileFoundation@5946 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Source/ETProtocolMirror.m
@@ -31,7 +31,7 @@
 	return [obj isMemberOfClass: [ETProtocolMirror class]] && 
 		[[obj name] isEqualToString: [self name]];
 }
-- (unsigned int) hash
+- (NSUInteger) hash
 {
 	return [[self name] hash];
 }
